+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.2.0
+
+- api changes
+- create docker project
+
+## 0.1.0
+
+- api changes
+- command line utility
+
 ## 0.0.1
 
 - Initial version